% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_compass.R
\name{compass_sinan}
\alias{compass_sinan}
\title{Sinan (Ancient Chinese Compass) Style (Simplified)}
\usage{
compass_sinan(
  line_col = "black",
  square_pad = 0.1,
  ring_outer = 0.35,
  ring_ratio = 0.65,
  labels = c("N", "E", "S", "W"),
  text_size = 12,
  text_face = "plain",
  text_family = "",
  text_col = "black",
  label_offset = 0.05,
  spoon_fill = "black",
  spoon_col = "black",
  spoon_scale = 0.8,
  inner_fill = "lightgrey",
  square_width = 2,
  outer_width = 2,
  inner_width = 1,
  spoon_width = 1
)
}
\value{
A grob representing a simplified "Sinan"-inspired compass.
}
\description{
Sinan (Ancient Chinese Compass) Style (Simplified)
}
