% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_loc.R
\name{geom_loc}
\alias{geom_loc}
\title{Add Spatial Point Data Layer with Color by Grouping}
\usage{
geom_loc(
  data,
  lon = NULL,
  lat = NULL,
  crs = "+proj=aeqd +lat_0=35 +lon_0=105 +ellps=WGS84 +units=m +no_defs",
  mapping = aes(),
  ...
)
}
\arguments{
\item{data}{A data frame, tibble, or \code{sf} object.}

\item{lon}{Character. The name of the longitude column in \code{data} (required if \code{data} is tabular).}

\item{lat}{Character. The name of the latitude column in \code{data} (required if \code{data} is tabular).}

\item{crs}{Character. Coordinate reference system (CRS) to which the data should be transformed. Defaults to "+proj=aeqd +lat_0=35 +lon_0=105 +ellps=WGS84 +units=m +no_defs".}

\item{mapping}{Aesthetic mappings created by \code{aes()}, such as \code{color}.}

\item{...}{Additional aesthetic mappings passed to \code{geom_sf}, such as \code{size}, \code{alpha}.}
}
\value{
A ggplot layer displaying spatial data as points in the specified CRS.
}
\description{
\code{geom_loc} is a flexible function that adds spatial data to a ggplot, supporting both \code{sf} and tabular data frames.
It allows color mapping based on a grouping variable in the data.
}
\examples{
# Example with tabular data from the package
data(pollen) # Load the example data provided by the package
ggplot() +
  geom_loc(
    data = pollen, lon = "Longitude", lat = "Latitude",
    mapping = aes(color = `Sample type`), size = 1, alpha = 0.7
  ) +
  labs(title = "Sample Points on Map")

}
