% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_compass.R
\name{north_arrow_classic}
\alias{north_arrow_classic}
\title{Classic North Arrow Style (Minimal)}
\usage{
north_arrow_classic(
  fill = c("white", "black"),
  line_col = "black",
  line_width = 2,
  text_col = "black",
  text_size = 12,
  text_face = "plain",
  text_family = ""
)
}
\value{
A grob representing a simple black arrow with "N".
}
\description{
Classic North Arrow Style (Minimal)
}
