% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmapcn-package.R
\docType{package}
\name{ggmapcn-package}
\alias{ggmapcn}
\alias{ggmapcn-package}
\title{ggmapcn: Customizable China Map Visualizations}
\description{
A 'ggplot2' extension for visualizing China’s map, offering customizable projections,
boundary styles, buffer zones, and annotation layers such as compass and scale bars.
}
\details{
Main features:
- Projection-aware compass (`annotation_compass`) and scale bar (`annotation_scalebar`).
- Built-in geodata management via `check_geodata`.
- Multiple compass styles (classic, rose, Sinan, guiding fish, etc.).
- Easy integration with ggplot2 and sf workflows.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rimagination.github.io/ggmapcn/}
  \item Report bugs at \url{https://github.com/Rimagination/ggmapcn/issues}
}

}
\author{
\strong{Maintainer}: Liang Ren \email{rl23@mails.tsinghua.edu.cn} (\href{https://orcid.org/0000-0002-2360-7900}{ORCID})

}
\keyword{package}
